requirements.txt
apache-airflow==2.10.5
dbt-core==1.9.4
dbt-postgres==1.9.0
pandas==2.2.3
requests==2.32.3
